Just Added SpringFox for Proper REST API Documentation in Swagger Format

+++ src/main/java/com/epam/training/stock/Stock.java
@@ -3,17 +3,18 @@ package com.epam.training.stock;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Stock {
     @JsonProperty("stock_id")
     private int sockId;
-    private float price;
+    private BigDecimal price;
     private int quantity;
 
-    public Stock(int sockId, float price, int quantity) {
+    public Stock(int sockId, BigDecimal price, int quantity) {
         this.sockId = sockId;
         this.price = price;
         this.quantity = quantity;
     }
-
 }
